switch to new collaborators component
