+++ packages/editor/api-report.md
@@ -619,7 +619,6 @@ export const debugFlags: {
     peopleMenu: Atom<boolean, unknown>;
     logMessages: Atom<never[], unknown>;
     resetConnectionEveryPing: Atom<boolean, unknown>;
-    newLiveCollaborators: Atom<boolean, unknown>;
 };
 
 // @internal (undocumented)
