+++ packages/editor/src/lib/components/Canvas.tsx
@@ -17,7 +17,6 @@ import { useQuickReactor } from '../hooks/useQuickReactor'
 import { useScreenBounds } from '../hooks/useScreenBounds'
 import { debugFlags } from '../utils/debug-flags'
 import { LiveCollaborators } from './LiveCollaborators'
-import { LiveCollaboratorsNext } from './LiveCollaboratorsNext'
 import { SelectionBg } from './SelectionBg'
 import { SelectionFg } from './SelectionFg'
 import { Shape } from './Shape'
@@ -128,11 +127,7 @@ export const Canvas = track(function Canvas({
 					<HintedShapeIndicator />
 					<SnapLinesWrapper />
 					<SelectionFg />
-					{debugFlags.newLiveCollaborators.value ? (
-						<LiveCollaboratorsNext />
-					) : (
-						<LiveCollaborators />
-					)}
+					<LiveCollaborators />
 				</div>
 			</div>
 		</div>
