+++ packages/editor/src/lib/utils/debug-flags.ts
@@ -23,7 +23,6 @@ export const debugFlags = {
 	peopleMenu: createDebugValue('tldrawPeopleMenu', false),
 	logMessages: createDebugValue('tldrawUiLog', []),
 	resetConnectionEveryPing: createDebugValue('tldrawResetConnectionEveryPing', false),
-	newLiveCollaborators: createDebugValue('tldrawNewLiveCollaborators', false),
 }
 
 declare global {
